+++ mne/preprocessing/maxfilter.py
@@ -7,15 +7,15 @@
 from ..externals.six import string_types
 import os
 from warnings import warn
-import logging
 
 import numpy as np
-from scipy import optimize, linalg
+from scipy import linalg
 
 from ..io import Raw
 from ..io.constants import FIFF
 from ..utils import logger, verbose
 from ..externals.six.moves import map
+from ..channels.utils import _fit_sphere
 
 
 @verbose
@@ -62,24 +62,7 @@ def fit_sphere_to_headshape(info, dig_kinds=(FIFF.FIFFV_POINT_EXTRA,),
 
     hsp = 1e3 * np.array(hsp)
 
-    # initial guess for center and radius
-    xradius = (np.max(hsp[:, 0]) - np.min(hsp[:, 0])) / 2
-    yradius = (np.max(hsp[:, 1]) - np.min(hsp[:, 1])) / 2
-
-    radius_init = (xradius + yradius) / 2
-    center_init = np.array([0.0, 0.0, np.max(hsp[:, 2]) - radius_init])
-
-    # optimization
-    x0 = np.r_[center_init, radius_init]
-    cost_fun = lambda x, hsp:\
-        np.sum((np.sqrt(np.sum((hsp - x[:3]) ** 2, axis=1)) - x[3]) ** 2)
-
-    disp = True if logger.level <= logging.INFO else False
-    x_opt = optimize.fmin_powell(cost_fun, x0, args=(hsp,), disp=disp)
-
-    origin_head = x_opt[:3]
-    radius = x_opt[3]
-
+    radius, origin_head = _fit_sphere(hsp)
     # compute origin in device coordinates
     trans = info['dev_head_t']
     if trans['from'] != FIFF.FIFFV_COORD_DEVICE\